%d/%D working decently

+++ libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: asyed <asyed@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/18 12:20:10 by asyed             #+#    #+#             */
-/*   Updated: 2017/11/14 02:09:09 by asyed            ###   ########.fr       */
+/*   Updated: 2017/11/15 13:31:21 by asyed            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *ptr, int val, size_t num)
 {
 	char	*d;
-	int		i;
+	size_t	i;
 
 	d = (char *)ptr;
 	i = 0;
